Added user and tracking tables,and index.html

+++ seed.py
@@ -3,6 +3,8 @@
 from model import Child, Location, connect_to_db, db
 from server import app
 
+from crud import create_user
+
 
 #__________________________________________________#
 
@@ -27,13 +29,13 @@ def get_children():
             print(lst)
             r1, r2, r3, r4, r5, r6, r7, r8, r9, r10, r11, r12, r13 = lst
 
-            db.session.add(Child(int(r1), r2, r3, r4, r5, int(r6), r7, r8, r9, r10, r11, r12, r13))
+            db.session.add(Child(r1, int(r2), r3, r4, r5, int(r6), r7, r8, r9, r10, r11, r12, r13))
             # db.session.add(Location(int(r1), r6, r7, r8)
-            db.session.add(Location(int(r1), r2, r3, r4, r5, float(r6), r7, r8, r9, r10, r11, r12, r13))
+            db.session.add(Location(r1, int(r2), r3, r4, r5, int(r6), r7, r8, r9, r10, r11, r12, r13))
 
     db.session.commit()
 
-
+# create_user()
 
 #__________________________________________________#
